requirements.txt
openai
openai-whisper
typer
pytube